refactor(worker): mirror URL structure under routes/<segment>.ts
Review feedback: route groups were scattered (auth/index.ts,
auth/identities.ts, tldraw-assets.ts, inline in worker.ts) with no
predictable URL-to-file mapping. Move every JSON sub-router under
`routes/<top-level-url-segment>.ts` so the URL always tells you
which file to open:

  routes/auth.ts            /auth/me, /auth/logout, /auth/identities*
  routes/workspaces.ts      /api/workspaces
  routes/documents.ts       /api/documents/* (CRUD + finalize +
                            snapshot + offline-cache + snapshot-status)
  routes/document-assets.ts /api/documents/:id/assets*
  routes/connect.ts         /api/connect/* (websocket)

Other moves:
- `worker.ts` is now entry-only: app construction, middleware mount
  points, the chained `routes` that captures `AppType`, the
  scheduled sweep, and the platform handler. No route definitions.
- `auth/index.ts` keeps only non-route helpers
  (`registerOAuthProviderRoutes`, `registerApiAuth`); the JSON routes
  it used to host (me/logout) moved into `routes/auth.ts`. The
  `auth/` directory remains a home for session/JWT/IdP helpers.
- `auth/identities.ts` deleted — folded into `routes/auth.ts`.
- Shared row-type + ownership-check helpers extracted to
  `route-helpers.ts` (used by both `documents.ts` and the schema-
  validation paths).
- `tldraw-assets.ts` keeps `assetRoutes` and its ~15 internal
  helpers together — moving the routes alone would have forced all
  the helpers public. `routes/document-assets.ts` is a one-line
  re-export shim so the URL→file rule still resolves; the cost is a
  single navigation hop.

No behavior change. Worker typecheck + 42 tests green; app
typecheck + 97 tests green.
